1-3_Data-Loader-Script/requirements.txt
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pandas>=2.1.3